+++ model/src/main/java/edu/java/exceptions/NotFoundException/ChatNotRegisteredException.java
@@ -1,15 +1,15 @@
 package edu.java.exceptions.NotFoundException;
 
 public class ChatNotRegisteredException extends NotFoundException {
-    public ChatNotRegisteredException(long id, String uri, String message) {
+    public ChatNotRegisteredException(Integer id, String uri, String message) {
         super(id, uri, message);
     }
 
-    public ChatNotRegisteredException(long id, String uri) {
+    public ChatNotRegisteredException(Integer id, String uri) {
         this(id, uri, DEFAULT_MESSAGE);
     }
 
-    public ChatNotRegisteredException(long id) {
+    public ChatNotRegisteredException(Integer id) {
         this(id, "", DEFAULT_MESSAGE);
     }
 }
